tests/wpt/web-platform-tests/content-security-policy/worker-src/service-fallback.https.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="worker-src https://{{host}}:{{ports[https][0]}}; child-src 'none'">
<script>
  var url = new URL("../support/ping.js", document.baseURI).toString();
  assert_service_worker_is_loaded(url, "Same-origin service worker allowed by host-source expression.");
</script>
